requirements.txt
Flask>=2.1.2
Werkzeug<2.1.0
gunicorn>=20.1.0
pyjwt>=2.4.0
requests>=2.27.1
cryptography>=3.4.8
flask-cors>=3.0.10
requests>=2.31.0
